requirements.txt
absl-py>=0.7.1
jmp>=0.0.2
numpy>=1.18.0
tabulate>=0.8.9
dm-haiku>=0.0.13
